requirements.txt
# Telegram Bot
aiogram==3.2.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0

# Logging
loguru==0.7.2

# Environment
python-dotenv==1.0.0

# Async operations
aiohttp>=3.8.0
aiofiles==23.2.1

# Google Sheets Integration
gspread==5.12.0
oauth2client==4.1.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
